requirements.txt
torch
transformers
pandas
numpy
pyyaml
regex
